+++ packages/antd-management-fast-framework/src/framework/DataSinglePageView/SinglePageDrawer/index.jsx
@@ -3,6 +3,7 @@ import classNames from 'classnames';
 import React from 'react';
 
 import {
+  checkObjectIsNullOrEmpty,
   checkStringIsNullOrWhiteSpace,
   isFunction,
   logCallTrack,
@@ -11,12 +12,14 @@ import {
 } from 'easy-soft-utility';
 
 import {
+  cardConfig,
   contentConfig,
   emptyLogic,
   listViewConfig,
   renderFurtherColorWhenNoCallProcess,
   renderFurtherPrefixWhenNoCallProcess,
 } from 'antd-management-fast-common';
+import { PageExtra } from 'antd-management-fast-component';
 
 import {
   DrawerExtra,
@@ -30,6 +33,8 @@ import { RefreshButton } from '../../DataListView/RefreshButton';
 import { SinglePage } from '../SinglePage';
 
 import styles from '../../DataListView/Drawer/index.less';
+
+const { ContentBox, SiderBox } = PageExtra;
 
 const primaryCallName = 'DataSinglePageView::SinglePageDrawer';
 
@@ -524,6 +529,7 @@ class SinglePageDrawer extends SinglePage {
             </Card>
           </div>
         </div>
+
         {this.renderPresetOther()}
       </div>
     );
@@ -662,6 +668,44 @@ class SinglePageDrawer extends SinglePage {
 
     const that = this;
 
+    const layoutConfig = this.establishPageContentLayoutConfig();
+    const siderConfig = this.establishPageContentLayoutSiderConfig();
+    const contentConfig = this.establishPageContentLayoutContentConfig();
+
+    layoutConfig.style = {
+      ...layoutConfig.style,
+      height: '100%',
+    };
+
+    siderConfig.style = {
+      ...siderConfig.style,
+      marginTop: '24px',
+      borderRadius: 0,
+    };
+
+    siderConfig.theme = 'light';
+
+    const siderTop = this.renderPresetSiderTopArea();
+    const siderBottom = this.renderPresetSiderBottomArea();
+
+    const siderBody =
+      checkObjectIsNullOrEmpty(siderTop) &&
+      checkObjectIsNullOrEmpty(siderBottom) ? null : (
+        <SiderBox
+          top={siderTop}
+          bottom={
+            checkObjectIsNullOrEmpty(siderTop)
+              ? siderBottom
+              : checkObjectIsNullOrEmpty(siderBottom)
+                ? this.buildCardCollectionArea({
+                    mode: cardConfig.wrapperType.page,
+                    list: [{}],
+                  })
+                : siderBottom
+          }
+        />
+      );
+
     return (
       <DrawerExtra
         flag={this.getVisibleFlag()}
@@ -694,7 +738,13 @@ class SinglePageDrawer extends SinglePage {
               height: '100%',
             }}
           >
-            {this.renderPresetDrawerInner()}
+            <ContentBox
+              layoutConfig={layoutConfig}
+              siderConfig={siderConfig}
+              siderBody={siderBody}
+              contentConfig={contentConfig}
+              contentBody={this.renderPresetDrawerInner()}
+            />
           </div>
         ) : null}
 
@@ -704,7 +754,13 @@ class SinglePageDrawer extends SinglePage {
               height: 'calc(100vh - 55px)',
             }}
           >
-            {this.renderPresetDrawerInner()}
+            <ContentBox
+              layoutConfig={layoutConfig}
+              siderConfig={siderConfig}
+              siderBody={siderBody}
+              contentConfig={contentConfig}
+              contentBody={this.renderPresetDrawerInner()}
+            />
           </div>
         ) : null}
 
@@ -714,7 +770,13 @@ class SinglePageDrawer extends SinglePage {
               height: 'calc(100vh - 55px)',
             }}
           >
-            {this.renderPresetDrawerInner()}
+            <ContentBox
+              layoutConfig={layoutConfig}
+              siderConfig={siderConfig}
+              siderBody={siderBody}
+              contentConfig={contentConfig}
+              contentBody={this.renderPresetDrawerInner()}
+            />
           </div>
         ) : null}
 
@@ -724,7 +786,13 @@ class SinglePageDrawer extends SinglePage {
               height: 'calc(100vh - 55px)',
             }}
           >
-            {this.renderPresetDrawerInner()}
+            <ContentBox
+              layoutConfig={layoutConfig}
+              siderConfig={siderConfig}
+              siderBody={siderBody}
+              contentConfig={contentConfig}
+              contentBody={this.renderPresetDrawerInner()}
+            />
           </div>
         ) : null}
       </DrawerExtra>