+++ examples/example-antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.9.124
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.7.85
+  - antd-management-fast-framework@2.12.96
+  - antd-management-fast-design-playground@1.0.241
+  - antd-management-fast-flow@1.0.169
+  - antd-management-fast-formily@1.0.180
+
 ## 1.9.123
 
 ### Patch Changes