+++ packages/antd-management-fast-component/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2.7.85
+
+### Patch Changes
+
+- update
+
 ## 2.7.84
 
 ### Patch Changes